rename media-player.[ch] to media-player-entry.[ch]
