+++ src/5gmsaf/provisioning-session.c
@@ -10,7 +10,7 @@ https://drive.google.com/file/d/1cinCiA778IErENZ3JN52VFW-1ffHpx7Z/view
 
 #include "provisioning-session.h"
 #include "application-server.h"
-#include "media-player.h"
+#include "media-player-entry.h"
 #include "context.h"
 #include "utilities.h"
